+++ osfmk/i386/hw_lock_types.h
@@ -90,7 +90,7 @@
  *	later in kern/lock.h..
  */
 struct hslock {
-	long	lock_data;
+	uintptr_t	lock_data;
 };
 typedef struct hslock hw_lock_data_t, *hw_lock_t;
 #define hw_lock_addr(hwl)	(&((hwl).lock_data))
